requirement.txt
streamlit>=1.25.0
pandas>=2.0
numpy>=1.26
scikit-learn>=1.3
scipy>=1.11
sentence-transformers>=2.2.2
torch>=2.10.0
faiss-cpu>=1.7.4.post2        # ou faiss-gpu si GPU
swifter>=1.2.0
